Refactor Go2 connector tests to use fixtures
Rename test file and refactor how external dependencies are mocked and imported. The module-level sys.modules mocking was replaced with a mock_external_modules fixture using monkeypatch, and a go2_module fixture now imports the connector after mocks are in place. default_config and custom_config fixtures were updated to use go2_module, global reset_mocks autouse was removed, and tests were adjusted to instantiate connector/config via go2_module. This improves test isolation and ensures external modules are mocked before importing the connector. The file was also renamed to test_move_game_controller_go2_game_controller.py.
